Fix: fix des problèmes cités

+++ API/services/activity_service.go
@@ -191,7 +191,7 @@ func GetDetailedActivitiesFromRange(from string, to string, idUser int) ([]*DTOs
 	return detailedActivitiesDTOs, err
 }
 
-func GetActivitiesCountFromCategoryId(categoryId string) (int, error) {
+func GetActivitiesCountFromCategoryId(categoryId string) (int64, error) {
 	activitiesCount, err := repositories.GetActivitiesCountFromCategoryId(categoryId)
 	if err != nil {
 		return 0, err
